Network.AWS.MarketplaceEntitlement

Description

AWS Marketplace Entitlement Service

This reference provides descriptions of the AWS Marketplace Entitlement Service API.

AWS Marketplace Entitlement Service is used to determine the entitlement of a customer to a given product. An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

Getting Entitlement Records

GetEntitlements - Gets the entitlements for a Marketplace product.

Service Configuration

marketplaceEntitlement :: Service

API version 2017-01-11 of the Amazon Marketplace Entitlement Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MarketplaceEntitlement.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters in your request was invalid.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to the GetEntitlements API are throttled.

InternalServiceErrorException

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred. Retry your request. If the problem persists, post a message with details on the AWS forums.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetEntitlements

module Network.AWS.MarketplaceEntitlement.GetEntitlements

Types

GetEntitlementFilterName

data GetEntitlementFilterName

Constructors

CustomerIdentifier

Dimension

Entitlement

data Entitlement

An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

See: entitlement smart constructor.

entitlement :: Entitlement

Creates a value of Entitlement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eDimension - The dimension for which the given entitlement applies. Dimensions represent categories of capacity in a product and are specified when the product is listed in AWS Marketplace.

eValue - The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

eExpirationDate - The expiration date represents the minimum date through which this entitlement is expected to remain valid. For contractual products listed on AWS Marketplace, the expiration date is the date at which the customer will renew or cancel their contract. Customers who are opting to renew their contract will still have entitlements with an expiration date.

eCustomerIdentifier - The customer identifier is a handle to each unique customer in an application. Customer identifiers are obtained through the ResolveCustomer operation in AWS Marketplace Metering Service.

eProductCode - The product code for which the given entitlement applies. Product codes are provided by AWS Marketplace when the product listing is created.

eDimension :: Lens' Entitlement (Maybe Text)

The dimension for which the given entitlement applies. Dimensions represent categories of capacity in a product and are specified when the product is listed in AWS Marketplace.

eValue :: Lens' Entitlement (Maybe EntitlementValue)

The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

eExpirationDate :: Lens' Entitlement (Maybe UTCTime)

The expiration date represents the minimum date through which this entitlement is expected to remain valid. For contractual products listed on AWS Marketplace, the expiration date is the date at which the customer will renew or cancel their contract. Customers who are opting to renew their contract will still have entitlements with an expiration date.

eCustomerIdentifier :: Lens' Entitlement (Maybe Text)

The customer identifier is a handle to each unique customer in an application. Customer identifiers are obtained through the ResolveCustomer operation in AWS Marketplace Metering Service.

eProductCode :: Lens' Entitlement (Maybe Text)

The product code for which the given entitlement applies. Product codes are provided by AWS Marketplace when the product listing is created.

EntitlementValue

data EntitlementValue

The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

See: entitlementValue smart constructor.

entitlementValue :: EntitlementValue

Creates a value of EntitlementValue with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evIntegerValue - The IntegerValue field will be populated with an integer value when the entitlement is an integer type. Otherwise, the field will not be set.

evDoubleValue - The DoubleValue field will be populated with a double value when the entitlement is a double type. Otherwise, the field will not be set.

evStringValue - The StringValue field will be populated with a string value when the entitlement is a string type. Otherwise, the field will not be set.

evBooleanValue - The BooleanValue field will be populated with a boolean value when the entitlement is a boolean type. Otherwise, the field will not be set.

evIntegerValue :: Lens' EntitlementValue (Maybe Int)

The IntegerValue field will be populated with an integer value when the entitlement is an integer type. Otherwise, the field will not be set.

evDoubleValue :: Lens' EntitlementValue (Maybe Double)

The DoubleValue field will be populated with a double value when the entitlement is a double type. Otherwise, the field will not be set.

evStringValue :: Lens' EntitlementValue (Maybe Text)

The StringValue field will be populated with a string value when the entitlement is a string type. Otherwise, the field will not be set.

evBooleanValue :: Lens' EntitlementValue (Maybe Bool)

The BooleanValue field will be populated with a boolean value when the entitlement is a boolean type. Otherwise, the field will not be set.